requirements.txt
langchain==0.0.335
neo4j-driver==5.12.0
openai==1.2.4
typing-inspect==0.8.0
typing_extensions==4.5.0
tiktoken==0.5.2
youtube-search==2.1.2